Add files via upload
Temperature Converter (Non-responsive)

This repository contains an implementation of a temperature converter using HTML, CSS, and JavaScript. While the converter provides functionality to convert temperatures between different units, please note that it is not currently responsive. It serves as a demonstration of my HTML, CSS, and JavaScript skills. Contributions and suggestions are welcome to enhance the responsiveness of the converter